backend/requirements.txt
python-telegram-bot==20.6
apscheduler==3.10.4
requests==2.31.0
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==21.2.0